factor out GIF generation
this fixes #17.
