+++ Fumen Utils_files/rendering/board-render-four.js
@@ -1,6 +1,5 @@
 import { boardSize } from "../global-utils.js";
-import { pageToBoard, renderBoardOnCanvas, getFumenMaxHeight } from "./board-render.js"
-import { GIFEncoder as gifenc, quantize, applyPalette } from 'https://unpkg.com/gifenc@1.0.3/dist/gifenc.esm.js';
+import { pageToBoard, renderBoardOnCanvas, getFumenMaxHeight, GenerateGIF } from "./board-render.js"
 
 function draw(fumenPage, numrows) {
 	var tileSize = Math.max(1, document.getElementById('cellSize').valueAsNumber || 0)
@@ -50,20 +49,6 @@ function drawFumens(fumenPages) {
 	return canvases
 }
 
-function GenerateFourGIF(canvases) {
-	let transparent = document.getElementById('transparency').checked
-	let delay = parseFloat(document.getElementById('delay').value)
-	const gif = new gifenc();
-	canvases.forEach(canvas => {
-		const { data, width, height } = canvas.getContext('2d').getImageData(0, 0, canvas.width, canvas.height)
-		const palette = quantize(data, 256, { format: 'rgba4444' });
-		const index = applyPalette(data, palette, { format: 'rgba4444' });
-		gif.writeFrame(index, width, height, { palette, transparent, delay }); //assumes that the first element in palette is [0,0,0,0].
-	})
-	gif.finish();
-	return gif;
-}
-
 function GIFDataURL(gif) {
 	let bytes = gif.stream.bytes()
 	return 'data:image/gif;base64,' + base64js.fromByteArray(bytes);
@@ -79,7 +64,7 @@ export default function fumencanvas(fumens) {
 			var data_url = canvas.toDataURL("image/png")
 		} else if (fumen.length >= 2) {
 			let canvases = drawFumens(fumen);
-			var data_url = GIFDataURL(GenerateFourGIF(canvases));
+			var data_url = GIFDataURL(GenerateGIF(canvases));
 		}
 
 		var img = new Image();
